KONIEC NOC

+++ Cwiczenia_1_APBD/Cwiczenia_1_APBD/Container.cs
@@ -31,7 +31,7 @@ public class Container
         this.LoadMass = 0;
     }
 
-    public virtual void LoadIn(int mass)
+    public virtual void LoadIn(double mass)
     {
         if (mass + this.LoadMass <= this.LoadMaxMass)
         {
